Update ax25.c
source_ssid and destination_ssid are both defined the same in config.h and in ax25.h.  Both are inserted to memory by adjusting a pointer. Therefore both should require moving the pointer back 7 bytes to place in correctly addressed position. I am puzzled that the header definition and the cast in this code line are both needed.